# Optimize .? in the FastRegexMatcher