This PR builds on top of #536 and optimize for the ".?" case. I don't feel strong about this optimization, but we believe it's not too risky (from a correctness perspective), may be another bullet in our optimization arsenal.

Performance change for the queries containing ".?":

```
goos: darwin
goarch: amd64
pkg: github.com/prometheus/prometheus/model/labels
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                              │ before-quest.txt │           after-quest.txt           │
                              │      sec/op      │   sec/op     vs base                │
FastRegexMatcher/fo.?-12             191.5n ± 1%   141.3n ± 1%  -26.21% (p=0.000 n=10)
FastRegexMatcher/foo.?-12            193.9n ± 3%   140.8n ± 2%  -27.40% (p=0.000 n=10)
FastRegexMatcher/f.?o-12             136.5n ± 2%   138.2n ± 2%   +1.28% (p=0.017 n=10)
FastRegexMatcher/.*foo.?-12         1843.0n ± 2%   298.2n ± 7%  -83.82% (p=0.000 n=10)
FastRegexMatcher/.?foo.+-12          850.9n ± 2%   289.7n ± 5%  -65.96% (p=0.000 n=10)
FastRegexMatcher/foo.?|bar-12        961.7n ± 2%   224.7n ± 2%  -76.63% (p=0.000 n=10)
geomean
```